requirements.txt
Werkzeug<3,>=0.16.0
Pillow>=6,<9
Flask
